Moved Windows specific files to a separate directory to keep a cleaner folder structure.
