+++ FItSNE.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "FItSNE", "src\FItSNE.vcxproj", "{D7401C22-FC52-4657-86D0-DBEBE8DCE006}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "FItSNE", "FItSNE.vcxproj", "{D7401C22-FC52-4657-86D0-DBEBE8DCE006}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
