+++ src/winlibs/stdafx.cpp
